+++ src/loop_optimize.h
@@ -3,8 +3,8 @@
 
 namespace simd_optimizations
 {
-	float calculateNormalLoop(const float* arr1, const float* arr2, int size);
-	float calculateOptimizedLoop(const float* arr1, const float* arr2, int size);
+	float calculateNormalLoop(const float* arr1, const float* arr2, size_t size);
+	float calculateOptimizedLoop(const float* arr1, const float* arr2, size_t size);
 }
 
 #endif //LOOP_OPTIMIZE_H_
